+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Loading model schema from database is now thread-safe. GH#28589
+
+    *Vikrant Chaudhary*
+
 *   Add `ActiveRecord::Base#cache_version` to support recyclable cache keys via the new versioned entries
     in `ActiveSupport::Cache`. This also means that `ActiveRecord::Base#cache_key` will now return a stable key
     that does not include a timestamp any more.